function/requirements.txt
feedparser
google-generativeai
google-cloud-texttospeech
pydub
google-cloud-storage
google-cloud-secret-manager
